Move testing function to new testing directory
